refactor(bcos-evm): consolidate the opstack module 14 file pairs -> 9
Pure file reorganization along cohesion lines; every function keeps its
name and body:

- OpBlock.{h,cpp} <- OpBlockExecute + OpBlockFinalize + OpBlockSeal
  (the processOpBlock -> finalizeOpBlock -> sealOpBlock pipeline in one
  translation unit; Finalize/Seal each had a single external consumer)
- OpReceipt.{h,cpp} <- OpReceiptMeta + OpReceiptEncode; OpDepositReceipt
  moves here from OpDepositTx.h so the receipt domain owns all receipt
  types (this also breaks the OpTransition<->OpReceipt include cycle the
  merge would otherwise create)
- OpTransition.{h,cpp} <- OpValidate + OpDepositTx (the validate/
  transition pairing constraint over props.fee is documented as a unit;
  runDeposit sits next to the runTxMessage core it shares)

Kept standalone: OpHost, OpForkSchedule (22 includers), OpFeeParams,
RollupCost, OpPredeploys, OpPrecompiles (predeploys vs precompiles stay
separate on purpose).

Verified: standalone bcos-evm rebuild green (fresh binaries);
BcosEvmEthTests + BcosEvmOpstackTests incl. the 33-vector op-geth
differential gate 100% pass.
