--- bcos-evm/bcos-evm/opstack/OpBlockExecute.h
@@ -1,48 +1,0 @@
-#pragma once
-
-#include <bcos-evm/opstack/OpDepositTx.h>
-#include <bcos-evm/opstack/OpForkSchedule.h>
-#include <bcos-evm/opstack/OpReceiptMeta.h>
-#include <functional>
-#include <span>
-#include <variant>
-#include <vector>
-
-namespace bcos::evm::opstack
-{
-/// One transaction within a block: deposit or normal tx (a normal tx must carry a signed envelope
-/// for L1 fee calculation).
-struct OpBlockTx
-{
-    std::variant<DepositTx, evmone::state::Transaction> tx;
-    evmc::bytes signedEnvelope;  // empty for deposit
-};
-
-/// Block execution result. receipts keep their original in-block order (M-B2's receipts-root /
-/// block-level bloom depend on this order; cumulative_gas_used is already filled in interleaved
-/// order).
-struct OpBlockResult
-{
-    std::vector<std::variant<OpDepositReceipt, OpTxReceipt>> receipts;
-    int64_t gasUsed = 0;                    // = last tx's cumulative
-    evmone::state::StateDiff finalizeDiff;  // end-of-block finalize output (already delivered via
-                                            // applyDiff)
-};
-
-/// Execute a whole block (spec §4.1 ordering): system_call_block_start → first L1 attributes
-/// deposit → loadOpFeeParams → per-transaction (gas pool / cumulative / per-transaction write-back)
-/// → finalizeOpBlock. Write-back callback: invoked immediately after each diff segment is produced;
-/// the view read by the next step must already reflect it.
-/// **discard-writes contract**: after any throw from this function, the caller must discard the
-/// entire write set already applied via applyDiff within this block (same semantics as op-geth
-/// Process discarding the whole statedb on error, state_processor.go:109-113). Throws
-/// std::runtime_error (block-level error): txs empty or first tx is not the L1 attributes deposit
-/// (to==OP_L1_BLOCK && from==OP_DEPOSITOR, stricter-than-spec); a deposit appears after a
-/// non-deposit (stricter-than-spec); any tx gasLimit exceeds the remaining block gas; is_system_tx;
-/// any validate error on a normal tx (op-geth has no failed-receipt mechanism for normal txs,
-/// state_processor.go:109-113).
-OpBlockResult processOpBlock(const evmone::state::StateView& view,
-    const evmone::state::BlockInfo& block, const evmone::state::BlockHashes& hashes,
-    std::span<const OpBlockTx> txs, const OpForkConfig& cfg, evmc::VM& vm, uint64_t chainId,
-    const std::function<void(const evmone::state::StateDiff&)>& applyDiff);
-}  // namespace bcos::evm::opstack
